dev-requirements.txt
#
# This file is autogenerated by pip-compile
# Make changes in dev-requirements.in, then run this to update:
#
#    pip-compile dev-requirements.in
#
astroid==1.4.4            # via pylint, pylint-plugin-utils
colorama==0.3.6           # via pylint
decorator==4.0.9          # via ipython, traitlets
factory-boy==2.7.0
fake-factory==0.5.5
ipython-genutils==0.1.0   # via traitlets
ipython==4.1.2
lazy-object-proxy==1.2.1  # via astroid
path.py==8.1.2            # via pickleshare
pexpect==4.0.1            # via ipython
pickleshare==0.6          # via ipython
ptyprocess==0.5.1         # via pexpect
pylint-django==0.7.1
pylint-plugin-utils==0.2.3  # via pylint-django
pylint==1.5.4
python-dateutil==2.5.0    # via fake-factory
simplegeneric==0.8.1      # via ipython
six==1.10.0               # via astroid, fake-factory, pylint, python-dateutil
traitlets==4.1.0          # via ipython
wrapt==1.10.6             # via astroid

# The following packages are commented out because they are
# considered to be unsafe in a requirements file:
# setuptools                # via ipython
